share existing translation logic & map access
